requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.8
django-appconf==1.0.5
django-compressor==2.4.1
django-environ==0.8.1
django-heroku==0.3.1
django-libsass==0.9
gunicorn==20.1.0
libsass==0.21.0
psycopg2==2.9.1
pytz==2021.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
